+++ aws-sdk-core/spec/aws/dynamodb/client_spec.rb
@@ -108,6 +108,37 @@ module Aws
         end
 
       end
+
+      describe '#enumerable responses' do
+
+        it 'can round trip paging params that contain item attributes' do
+          client = Client.new(stub_responses: {
+            query: {
+              :items=> [
+                {"indexhash"=>"hash", "id"=>"id-1", "indexrange"=>"range-1"},
+                {"indexhash"=>"hash", "id"=>"id-2", "indexrange"=>"range-2"},
+              ],
+              :count=>2,
+              :scanned_count=>2,
+              :last_evaluated_key=>{
+                "indexhash"=>"hash", "id"=>"id-160", "indexrange"=>"range-160"
+              }
+            }
+          })
+          result = client.query(
+            table_name: 'table-name',
+            limit: 2,
+            index_name: 'indexname',
+            key_condition_expression: 'indexhash = :value1',
+            expression_attribute_values: { ':value1' => 'hash'}
+          )
+          result2 = result.next_page
+          expect(result2.context.params[:expression_attribute_values]).to eq(
+            ':value1' => { s:'hash'}
+          )
+        end
+
+      end
     end
   end
 end